move Histogram and Regression to ScottPlot.Statistics namespace
